refactor: extract ApplePayModuleUtilitiesTests into own file (COSDK-473)
