+++ example/ios/AdyenExampleTests/Modules/ApplePayModuleTests.swift
@@ -272,118 +272,6 @@ final class ApplePayModuleTests: XCTestCase {
     }
 }
 
-final class ApplePayModuleUtilitiesTests: XCTestCase {
-
-    // MARK: - PKShippingMethod.jsonObject
-
-    func test_pkShippingMethod_jsonObject_containsRequiredFields() {
-        // GIVEN
-        let method = PKShippingMethod(label: "Standard", amount: 5.00)
-        method.identifier = "standard"
-        method.detail = "5–7 days"
-
-        // WHEN
-        let json = method.jsonObject
-
-        // THEN
-        XCTAssertEqual(json["label"] as? String, "Standard")
-        XCTAssertEqual(json["amount"] as? String, "5")
-        XCTAssertEqual(json["identifier"] as? String, "standard")
-        XCTAssertEqual(json["detail"] as? String, "5–7 days")
-    }
-
-    func test_pkShippingMethod_jsonObject_typeIsPending() {
-        // GIVEN
-        let method = PKShippingMethod(label: "TBD", amount: 0, type: .pending)
-
-        // WHEN
-        let json = method.jsonObject
-
-        // THEN
-        XCTAssertEqual(json["type"] as? String, "pending")
-    }
-
-    func test_pkShippingMethod_jsonObject_typeIsFinal() {
-        // GIVEN
-        let method = PKShippingMethod(label: "Express", amount: 15, type: .final)
-
-        // WHEN
-        let json = method.jsonObject
-
-        // THEN
-        XCTAssertEqual(json["type"] as? String, "final")
-    }
-
-    func test_pkShippingMethod_jsonObject_omitsNilIdentifierAndDetail() {
-        // GIVEN
-        let method = PKShippingMethod(label: "Free", amount: 0)
-
-        // WHEN
-        let json = method.jsonObject
-
-        // THEN
-        XCTAssertNil(json["identifier"])
-        XCTAssertNil(json["detail"])
-    }
-
-    // MARK: - applePayError(from:)
-
-    func test_applePayError_returnsNil_forUnknownType() {
-        XCTAssertNil(applePayError(from: ["type": "unknown", "message": "error"]))
-    }
-
-    func test_applePayError_returnsNil_whenMissingMessage() {
-        XCTAssertNil(applePayError(from: ["type": "shippingAddress"]))
-    }
-
-    func test_applePayError_shippingAddress_returnsError() {
-        let error = applePayError(from: ["type": "shippingAddress", "field": "postalCode", "message": "Invalid postal code"])
-        XCTAssertNotNil(error)
-    }
-
-    func test_applePayError_billingAddress_returnsError() {
-        let error = applePayError(from: ["type": "billingAddress", "field": "city", "message": "Invalid city"])
-        XCTAssertNotNil(error)
-    }
-
-    func test_applePayError_contactField_returnsError() {
-        let error = applePayError(from: ["type": "contactField", "field": "phoneNumber", "message": "Invalid phone"])
-        XCTAssertNotNil(error)
-    }
-
-    func test_applePayError_shippingAddress_usesDefaultKey_whenFieldIsNil() {
-        let error = applePayError(from: ["type": "shippingAddress", "message": "Error"])
-        XCTAssertNotNil(error)
-    }
-
-    // MARK: - ApplePayPaymentMethod.supportedNetworks
-
-    func test_supportedNetworks_returnsAllAvailable_whenBrandsIsNil() throws {
-        // GIVEN
-        let dict: NSDictionary = ["type": "applepay", "name": "Apple Pay"]
-        let method: ApplePayPaymentMethod = try dict.decode()
-
-        // WHEN
-        let networks = method.supportedNetworks
-
-        // THEN
-        XCTAssertFalse(networks.isEmpty)
-    }
-
-    func test_supportedNetworks_filtersToMatchingBrands() throws {
-        // GIVEN
-        let dict: NSDictionary = ["type": "applepay", "name": "Apple Pay", "brands": ["visa"]]
-        let method: ApplePayPaymentMethod = try dict.decode()
-
-        // WHEN
-        let networks = method.supportedNetworks
-
-        // THEN
-        XCTAssertTrue(networks.contains(.visa))
-        XCTAssertFalse(networks.contains(.masterCard))
-    }
-}
-
 private class TestableApplePayModule: ApplePayModule {
     var dismissCalled = false
     var dismissSuccessValue = false
